Update slickDraggableGrouping.ts
Having this reorderedColumns local to the whole functions makes it outdated at many cases. We are trying to set the columns to the initial value which is wrong. Also this field should be local to the end event.

The bug we observed is getting error for missing columns in the reorderedColumns on drag/drop column