fix: update slickRowDetailView.ts
This Pull Request adds a check to ensure item exists before trying to access a property in it.

In situations where rows need to be updated like changing the page, filtering, sorting, etc. the event `onBeforeRemoveCachedRow` is triggered which in turn triggers the RowDetailViewPlugin's method handleRemoveRow.
Sometimes the index provided to this method can't be found in the dataset so the line
`const item = this.dataView.getItemByIdx(rowIndex)` yields undefined.